requirements.txt
certifi==2020.12.5
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.22
protobuf==3.14.0
PyYAML==5.3.1
six==1.15.0
visitor==0.1.3
Werkzeug==1.0.1
wincertstore==0.2
